requirements.txt
ping3
pyyaml
pymysql
pandas
matplotlib
pytelegrambotapi